+++ src/gpu_api.cpp
@@ -69,8 +69,8 @@ std::vector<Reference> create_sketches_cuda(const std::string& db_name,
 		// memory for filter and nthash only need to be allocated once
 		copyNtHashTablesToDevice();
 		GPUCountMin countmin_filter();
-        if (min_count > std::numeric_limits<uint8_t>::max()) {
-            min_count = std::numeric_limits<uint8_t>::max();
+        if (min_count > std::numeric_limits<unsigned int>::max()) {
+            min_count = std::numeric_limits<unsigned int>::max();
         }
 
 		for (int i = 0; i < files.size(); i++) {
